# crimson/os/seastore: make JournalSubmitter aware of header merging