This PR starts at commit "crimson/os/seastore: add debug logs to print transaction name" and implements TODOs from https://github.com/ceph/ceph/pull/44041

1. As a result, this PR with #44041 together reduces writes for 4KB-aligned padding from ratio ~0.65 to ~0.2, reducing 70% writes from metadata padding:
  1.1. before: ![before](https://user-images.githubusercontent.com/7736006/143805928-dc121e48-78ec-477a-b844-52754f869906.png)
  1.2. after: ![after_new_1](https://user-images.githubusercontent.com/7736006/143805942-395688b9-7445-4860-ab84-c5229f3521e7.png)

2. This feature still allows both batching and concurrent io when submitting journals:
  2.1. before: ![before_2](https://user-images.githubusercontent.com/7736006/143811519-eafe7197-1823-4cb3-b231-111aafbda307.png)
  2.2. after: ![after_new_2](https://user-images.githubusercontent.com/7736006/143811541-bf44240b-ab4a-4608-b17a-dfd7adc9c5b4.png)

3. From overall write amplification analysis, this feature reduces the gap (overhead) between the device-level write and the cache-level data commits:
  3.1. before (see SEGMENTED_WRITE vs VALID_EXTENT_WRITE): ![before_4](https://user-images.githubusercontent.com/7736006/143812157-bc6cfa1f-18b1-404a-97f5-9cceb0d4f850.png)
  3.2. after (see SEGMENTED_WRITE vs AW_VALID_DATA): ![after_new_4](https://user-images.githubusercontent.com/7736006/143812182-6d6fc4cf-0ec6-411a-913f-ac66383caa1a.png)

4. The final improvements to rados bench throughput/latency is minor, I think there are reasons:
  4.1. The test workload makes SeaStore to be CPU-bound instead of IO-bound, because the reactor utilization reaches ~100%: ![reactorutil](https://user-images.githubusercontent.com/7736006/143813013-36d591c7-0c61-4316-ab18-8150d770bd0b.png)
  4.2. Cleaning overhead with the side-effects are too significant, see 3.2 AW_VALID_DATA vs MUTATE_TRANS_DATA_WRITE;
  4.3. Device read overhead is too significant due to lacking cache LRU, see 3.2 SEGMENTED_READ vs SEGMENTED_WRITE;

<!--
  - Please give your pull request a title like

      [component]: [short description]

  - Please use this format for each git commit message:

      [component]: [short description]

      [A longer multiline description]

      Fixes: [ticket URL on tracker.ceph.com, create one if necessary]
      Signed-off-by: [Your Name] <[your email]>

    For examples, use "git log".

  - The Signed-off-by line in every git commit is important; see <span class="x x-first x-last">[Submitting Patches to Ceph](https://github.com/ceph/ceph/blob/master/</span>SubmittingPatches.rst<span class="x x-first x-last">)</span>
-->

## Checklist
- Tracker (select at least one)
  - [ ] References tracker ticket
  - [ ] Very recent bug; references commit where it was introduced
  - [x] New feature (ticket optional)
  - [ ] Doc update (no ticket needed)
  - [ ] Code cleanup (no ticket needed)
- Component impact
  - [ ] Affects [Dashboard](https://tracker.ceph.com/projects/dashboard/issues/new), opened tracker ticket
  - [ ] Affects [Orchestrator](https://tracker.ceph.com/projects/orchestrator/issues/new), opened tracker ticket
  - [x] No impact that needs to be tracked
- Documentation (select at least one)
  - [ ] Updates relevant documentation
  - [x] No doc update is appropriate
- Tests (select at least one)
  - [ ] Includes [unit test(s)](https://docs.ceph.com/en/latest/dev/developer_guide/tests-unit-tests/)
  - [ ] Includes [integration test(s)](https://docs.ceph.com/en/latest/dev/developer_guide/testing_integration_tests/)
  - [ ] Includes bug reproducer
  - [x] No tests

<details>
<summary>Show available Jenkins commands</summary>

- `jenkins retest this please`
- `jenkins test classic perf`
- `jenkins test crimson perf`
- `jenkins test signed`
- `jenkins test make check`
- `jenkins test make check arm64`
- `jenkins test submodules`
- `jenkins test dashboard`
- `jenkins test dashboard cephadm`
- `jenkins test api`
- `jenkins test docs`
- `jenkins render docs`
- `jenkins test ceph-volume all`
- `jenkins test ceph-volume tox`

</details>