@cyx1231st 2021-11-30
Looks some tests went into a loop to do journal trimming endlessly, I guess something was wrong with the empty record optimization.