@cyx1231st 2021-12-03
> 4. The final improvements to rados bench throughput/latency is minor, I think there are reasons:
>     4.1. The test workload makes SeaStore to be CPU-bound instead of IO-bound ....
>     ...

@liu-chunmei FYI, this is the rados bench results with 30s tests, with and without the two PRs:

After-throughput (MB/sec):
| threads | round1 | round2 | round3 |
| ---| --- | --- | --- |
| 1 | 4.81 | 4.66 | 5.00 |
| 8 | 13.01 | 13.11 | 13.41 |
| 32 | 14.73 | 14.82 | 15.41 |

Before-throughput (MB/sec):
| threads | round1 | round2 | round3 |
| ---| --- | --- | --- |
| 1 | 4.91 | 4.96 | 4.71 |
| 8 | 12.38 | 13.29 | 12.79 |
| 32 | 15.12 | 14.94 | 14.74 |

After-latency (ms):
| threads | round1 | round2 | round3 |
| ---| --- | --- | --- |
| 1 | 0.81 | 0.83 | 0.78 |
| 8 | 2.40 | 2.38 |  2.33 |
| 32 | 8.48 | 8.43 | 8.11 |

Before-latency (ms):
| threads | round1 | round2 | round3 |
| ---| --- | --- | --- |
| 1 | 0.79 | 0.78 | 0.83 |
| 8 | 2.52 | 2.35 | 2.44 |
| 32 | 8.26 | 8.36 | 8.47 |

This is why I concluded that the user-end performance is not changed, statistically. (I haven't take much time to tune the related options, though)