@cyx1231st 2021-12-09
Hit a new issue, but seems independent to this PR.
https://tracker.ceph.com/issues/53555

Also splitted out empty transaction optimization to https://github.com/ceph/ceph/pull/44256